+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
       android:theme="@style/AppTheme"
       >
     <activity
-        android:name=".MainActivity"
+        android:name=".list.MainActivity"
         android:theme="@style/noAnimTheme"
         >
       <intent-filter>
