+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.smassive.bottombarmvp.MainActivity"
+    tools:context="com.smassive.bottombarmvp.list.MainActivity"
     >
 
   <android.support.v7.widget.RecyclerView
